(RES): move `declarations` to RRE
